+++ CHANGES.md
@@ -2,7 +2,7 @@
 
 ## (unreleased)
 
-- _nothing_
+- Ignore weblate user on contributing agreement check (2nd chanche) [erral, fredvd]_
 
 ## 1.2.9 (2025-03-31)